handiwork/requirements/base.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.3
django-allauth==0.44.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-phonenumber-field==5.1.0
django-toolbelt==0.0.1
future==0.16.0
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.0
phonenumberslite==8.12.23
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
typing-extensions==3.10.0.0
urllib3==1.26.4
